+++ engine/runtime/src/main/java/org/enso/compiler/TreeToIr.java
@@ -1,22 +1,23 @@
 package org.enso.compiler;
 
-import java.util.ArrayList;
-
 import org.enso.compiler.core.IR;
+import org.enso.compiler.core.IR$Application$Literal$Sequence;
 import org.enso.compiler.core.IR$Application$Operator$Binary;
 import org.enso.compiler.core.IR$Application$Prefix;
 import org.enso.compiler.core.IR$CallArgument$Specified;
+import org.enso.compiler.core.IR$Case$Branch;
+import org.enso.compiler.core.IR$Case$Expr;
 import org.enso.compiler.core.IR$Comment$Documentation;
 import org.enso.compiler.core.IR$DefinitionArgument$Specified;
 import org.enso.compiler.core.IR$Error$Syntax;
 import org.enso.compiler.core.IR$Error$Syntax$InvalidBaseInDecimalLiteral$;
-import org.enso.compiler.core.IR$Error$Syntax$InvalidImport$;
 import org.enso.compiler.core.IR$Error$Syntax$UnexpectedDeclarationInType$;
 import org.enso.compiler.core.IR$Error$Syntax$UnexpectedExpression$;
 import org.enso.compiler.core.IR$Error$Syntax$UnsupportedSyntax;
 import org.enso.compiler.core.IR$Expression$Binding;
 import org.enso.compiler.core.IR$Expression$Block;
 import org.enso.compiler.core.IR$Function$Lambda;
+import org.enso.compiler.core.IR$Function$Binding;
 import org.enso.compiler.core.IR$Literal$Text;
 import org.enso.compiler.core.IR$Literal$Number;
 import org.enso.compiler.core.IR$Module$Scope$Definition;
@@ -33,6 +34,9 @@ import org.enso.compiler.core.IR$Name$Blank;
 import org.enso.compiler.core.IR$Name$Literal;
 import org.enso.compiler.core.IR$Name$MethodReference;
 import org.enso.compiler.core.IR$Name$Qualified;
+import org.enso.compiler.core.IR$Pattern$Constructor;
+import org.enso.compiler.core.IR$Pattern$Name;
+import org.enso.compiler.core.IR$Pattern$Literal;
 import org.enso.compiler.core.IR$Type$Ascription;
 import org.enso.compiler.core.IR$Type$Function;
 import org.enso.compiler.core.IR.IdentifiedLocation;
@@ -41,6 +45,7 @@ import org.enso.compiler.core.ir.MetadataStorage;
 import org.enso.compiler.exception.UnhandledEntity;
 import org.enso.syntax.text.Location;
 import org.enso.syntax2.ArgumentDefinition;
+import org.enso.syntax2.Case;
 import org.enso.syntax2.Either;
 import org.enso.syntax2.FractionalDigits;
 import org.enso.syntax2.Line;
@@ -51,6 +56,7 @@ import org.enso.syntax2.Tree;
 
 import scala.Option;
 import scala.collection.immutable.List;
+import scala.jdk.javaapi.CollectionConverters;
 
 final class TreeToIr {
   static final TreeToIr MODULE = new TreeToIr();
@@ -82,20 +88,32 @@ final class TreeToIr {
         for (Line line : b.getStatements()) {
           switch (line.getExpression()) {
             case Tree.Assignment a -> {
-              var name = switch (a.getPattern()) {
-                case Tree.Ident id -> new IR$Name$Literal(
-                  id.getToken().codeRepr(), false,
-                  getIdentifiedLocation(id),
-                  meta(), diag()
-                );
+              var r = switch (a.getPattern()) {
+                case Tree.Ident id -> {
+                  var name = new IR$Name$Literal(
+                    id.getToken().codeRepr(), false,
+                    getIdentifiedLocation(id),
+                    meta(), diag()
+                  );
+                  yield new IR$Name$MethodReference(
+                    Option.empty(),
+                    name,
+                    name.location(),
+                    meta(), diag()
+                  );
+                }
+                case Tree.OprApp opr -> {
+                  var pointer = buildQualifiedName(opr.getLhs());
+                  var name = buildName(opr.getRhs());
+                  yield new IR$Name$MethodReference(
+                    Option.apply(pointer),
+                    name,
+                    Option.empty(),
+                    meta(), diag()
+                  );
+                }
                 default -> throw new IllegalStateException("Not an identifier: " + a.getPattern());
               };
-              var r = new IR$Name$MethodReference(
-                Option.empty(),
-                name,
-                name.location(),
-                meta(), diag()
-              );
               var m = new IR$Module$Scope$Definition$Method$Binding(
                 r,
                 nil(),
@@ -127,6 +145,10 @@ final class TreeToIr {
 //            }
             case Tree.TypeSignature def -> {
               var t = translateModuleSymbol(def);
+              bindings = cons(t, bindings);
+            }
+            case Tree.ArgumentBlockApplication app -> {
+              var t = translateComment(app);
               bindings = cons(t, bindings);
             }
             case null -> {
@@ -305,7 +327,7 @@ final class TreeToIr {
           getIdentifiedLocation(fn),
           meta(), diag()
         );
-        var args = translateArgumentsDefs(fn.getArgs());
+        var args = translateArgumentsDefinition(fn.getArgs());
         var body = translateExpression(fn.getBody(), false);
 
         yield new IR$Module$Scope$Definition$Method$Binding(
@@ -386,22 +408,8 @@ final class TreeToIr {
     };
   }
 
-  private List<IR.DefinitionArgument> translateArgumentsDefs(java.util.List<ArgumentDefinition> args) {
-    ArrayList<Tree> params = new ArrayList<>();
-    for (var d : args) {
-      params.add(d.getPattern());
-    }
-    return translateArgumentsDefinition(params);
-  }
-  private List<IR.DefinitionArgument> translateArgumentsDefinition(java.util.List<Tree> params) {
-      return translateArgumentsDefinition(params, nil());
-  }
-  private List<IR.DefinitionArgument> translateArgumentsDefinition(java.util.List<Tree> params, List<IR.DefinitionArgument> args) {
-      for (var p : params) {
-        var m = translateArgumentDefinition(p, false);
-        args = cons(m, args);
-      }
-    return args.reverse();
+  private List<IR.DefinitionArgument> translateArgumentsDefinition(java.util.List<ArgumentDefinition> args) {
+    return CollectionConverters.asScala(args.stream().map(p -> translateArgumentDefinition(p)).iterator()).toList();
   }
 
   /** Translates the body of a type expression.
@@ -410,14 +418,8 @@ final class TreeToIr {
     * @return the [[IR]] representation of `body`
     */
   private scala.collection.immutable.List<IR> translateTypeBody(java.util.List<Line> block, boolean found) {
-    List<IR> result = nil();
-    for (var line : block) {
-      var expr = translateTypeBodyExpression(line);
-      if (expr != null) {
-        result = cons(expr, result);
-      }
-    }
-    return result.reverse();
+    var ir = block.stream().map(line -> translateTypeBodyExpression(line)).filter(line -> line != null).iterator();
+    return CollectionConverters.asScala(ir).toList();
   }
 
   /** Translates any expression that can be found in the body of a type
@@ -587,23 +589,65 @@ final class TreeToIr {
             );
             yield prefix;
           }
+          case "->" -> {
+            if (insideTypeSignature) {
+              var literal = translateExpression(app.getLhs(), insideTypeSignature);
+              var body = translateExpression(app.getRhs(), insideTypeSignature);
+              var args = switch (body) {
+                case IR$Type$Function fn -> {
+                  body = fn.result();
+                  yield cons(literal, fn.args());
+                }
+                default -> cons(literal, nil());
+              };
+              yield new IR$Type$Function(args, body, Option.empty(), meta(), diag());
+            } else {
+              // Old-style lambdas; this syntax will be eliminated after the parser transition is complete.
+              var arg = app.getLhs();
+              var isSuspended = false;
+              if (arg instanceof Tree.UnaryOprApp susApp && "~".equals(susApp.getOpr().codeRepr())) {
+                  arg = susApp.getRhs();
+                  isSuspended = true;
+              }
+              IR.Name name = switch (arg) {
+                case Tree.Wildcard wild -> new IR$Name$Blank(getIdentifiedLocation(wild.getToken()), meta(), diag());
+                case Tree.Ident id -> {
+                  IR.Expression identifier = translateIdent(id, false);
+                  yield switch (identifier) {
+                    case IR.Name name_ -> name_;
+                    default -> throw new UnhandledEntity(identifier, "translateExpression");
+                  };
+                }
+                default -> throw new UnhandledEntity(arg, "translateExpressiontranslateArgumentDefinition");
+              };
+              var arg_ = new IR$DefinitionArgument$Specified(
+                      name,
+                      Option.empty(),
+                      Option.empty(),
+                      isSuspended,
+                      getIdentifiedLocation(arg),
+                      meta(),
+                      diag()
+              );
+              List<IR.DefinitionArgument> args = cons(arg_, nil());
+              var body = translateExpression(app.getRhs(), false);
+              yield new IR$Function$Lambda(args, body, getIdentifiedLocation(tree), true, meta(), diag());
+            }
+          }
           default -> {
-            if (op.getProperties() != null) {
-              var lhs = translateCallArgument(app.getLhs(), insideTypeSignature);
-              var rhs = translateCallArgument(app.getRhs(), insideTypeSignature);
-              yield new IR$Application$Operator$Binary(
-                lhs,
-                new IR$Name$Literal(
-                  op.codeRepr(), true,
-                  getIdentifiedLocation(app),
-                  meta(), diag()
-                ),
-                rhs,
+            var lhs = translateCallArgument(app.getLhs(), insideTypeSignature);
+            var rhs = translateCallArgument(app.getRhs(), insideTypeSignature);
+            yield new IR$Application$Operator$Binary(
+              lhs,
+              new IR$Name$Literal(
+                op.codeRepr(), true,
                 getIdentifiedLocation(app),
                 meta(), diag()
-              );
-            }
-            throw new UnhandledEntity(tree, op.codeRepr());
+              ),
+              rhs,
+              getIdentifiedLocation(app),
+              meta(), diag()
+            );
           }
         };
       }
@@ -612,7 +656,26 @@ final class TreeToIr {
         var fn = translateExpression(app.getFunc(), cons(app.getArg(), moreArgs), insideTypeSignature, false);
         yield fn;
       }
-      case Tree.Number n -> translateDecimalLiteral(n, n.getInteger(), n.getFractionalDigits());
+      case Tree.NamedApp app -> {
+        var fn = translateExpression(app.getFunc(), cons(app, moreArgs), insideTypeSignature, false);
+        yield fn;
+      }
+      case Tree.Array arr -> {
+        List<IR.Expression> items = nil();
+        if (arr.getFirst() != null) {
+          var exp = translateExpression(arr.getFirst(), insideTypeSignature);
+          items = cons(exp, items);
+        }
+        for (var next : arr.getRest()) {
+          var exp = translateExpression(next.getBody(), insideTypeSignature);
+          items = cons(exp, items);
+        }
+        yield new IR$Application$Literal$Sequence(
+          items.reverse(),
+          getIdentifiedLocation(arr), meta(), diag()
+        );
+      }
+      case Tree.Number n -> translateDecimalLiteral(n);
       case Tree.Ident id -> {
         var exprId = translateIdent(id, isMethod);
         if (moreArgs.isEmpty()) {
@@ -696,37 +759,63 @@ final class TreeToIr {
       case Tree.Group group -> {
         yield translateExpression(group.getBody(), moreArgs, insideTypeSignature, isMethod);
       }
-      case Tree.TextLiteral txt -> {
-        var fullTxt = txt.codeRepr().trim();
-        var t = fullTxt.substring(1, fullTxt.length() - 1);
-        yield new IR$Literal$Text(t, getIdentifiedLocation(txt), meta(), diag());
-      }
-      case Tree.Arrow arrow when insideTypeSignature -> {
-        List<IR.Expression> args = nil();
-        Tree treeBody = null;
-        {
-          Tree.Arrow head = arrow;
-          while (head != null) {
-            for (var a : head.getArguments()) {
-              var literal = buildName(a);
-              args = cons(literal, args);
-            }
-            treeBody = head.getBody();
-            head = switch (treeBody) {
-              case Tree.Arrow a -> a;
-              default -> null;
-            };
+      case Tree.TextLiteral txt -> translateLiteral(txt);
+      case Tree.CaseOf cas -> {
+        var expr = translateExpression(cas.getExpression(), insideTypeSignature);
+        List<IR$Case$Branch> branches = nil();
+        for (var line : cas.getCases()) {
+          if (line.getCase() == null) {
+            continue;
           }
-          args = args.reverse();
+          var br = translateCaseBranch(line.getCase());
+          branches = cons(br, branches);
         }
-        var body = translateExpression(treeBody, isMethod);
-        yield new IR$Type$Function(args, body, Option.empty(), meta(), diag());
+        yield new IR$Case$Expr(expr, branches.reverse(), getIdentifiedLocation(tree), meta(), diag());
       }
+      case Tree.Function fun -> {
+        var name = buildName(fun.getName());
+        var args = translateArgumentsDefinition(fun.getArgs());
+        var body = translateExpression(fun.getBody(), false);
 
-      case Tree.Arrow arrow -> {
-        var arguments = translateArgumentsDefinition(arrow.getArguments());
-        var body = translateExpression(arrow.getBody(), isMethod);
-        yield new IR$Function$Lambda(arguments, body, getIdentifiedLocation(arrow), true, meta(), diag());
+        yield new IR$Function$Binding(name, args, body,
+            getIdentifiedLocation(fun), true, meta(), diag()
+        );
+      }
+      case Tree.OprSectionBoundary bound -> {
+        var ast = translateExpression(bound.getAst(), insideTypeSignature);
+        yield switch (ast) {
+          case IR$Application$Prefix p -> p.function();
+          default -> ast;
+        };
+      }
+      case Tree.UnaryOprApp un -> switch (translateExpression(un.getRhs(), insideTypeSignature)) {
+        case IR$Literal$Number n when "-".equals(un.getOpr().codeRepr()) -> n.copy(
+          n.copy$default$1(),
+          "-" + n.copy$default$2(),
+          n.copy$default$3(),
+          n.copy$default$4(),
+          n.copy$default$5(),
+          n.copy$default$6()
+        );
+        case IR.Expression e -> e;
+      };
+      case Tree.TypeSignature sig -> {
+        var methodName = buildName(sig, sig.getVariable());
+        var methodReference = new IR$CallArgument$Specified(
+                Option.empty(),
+                methodName,
+                getIdentifiedLocation(sig),
+                meta(), diag()
+        );
+        var opName = buildName(null, sig.getOperator(), true);
+        var signature = translateCallArgument(sig.getType(), true);
+        yield new IR$Application$Operator$Binary(methodReference, opName, signature, getIdentifiedLocation(sig), meta(), diag());
+      }
+      case Tree.TemplateFunction templ -> {
+        yield translateExpression(templ.getAst(), insideTypeSignature);
+      }
+      case Tree.Wildcard wild -> {
+        yield new IR$Name$Blank(getIdentifiedLocation(wild), meta(), diag());
       }
       default -> throw new UnhandledEntity(tree, "translateExpression");
     };
@@ -735,8 +824,6 @@ final class TreeToIr {
       .unapply(maybeParensedInput)
       .getOrElse(maybeParensedInput)
 
-    inputAst match {
-      case AST.Def(consName, _, _) =>
         IR.Error
           .Syntax(inputAst, IR.Error.Syntax.TypeDefinedInline(consName.name))
       case AstView.UnaryMinus(expression) =>
@@ -860,6 +947,10 @@ final class TreeToIr {
     */
   }
 
+  IR.Expression translateDecimalLiteral(Tree.Number ast) {
+    return translateDecimalLiteral(ast, ast.getInteger(), ast.getFractionalDigits());
+  }
+
   IR.Expression translateDecimalLiteral(
     Tree ast,
     Token.Digits intPart,
@@ -890,6 +981,12 @@ final class TreeToIr {
         getIdentifiedLocation(ast), meta(), diag()
       );
     }
+  }
+
+  IR.Literal translateLiteral(Tree.TextLiteral txt) {
+    var fullTxt = txt.codeRepr().trim();
+    var t = fullTxt.substring(1, fullTxt.length() - 1);
+    return new IR$Literal$Text(t, getIdentifiedLocation(txt), meta(), diag());
   }
 
   /** Translates a program literal from its [[AST]] representation into
@@ -1061,134 +1158,40 @@ final class TreeToIr {
 
 
   /** Translates an argument definition from [[AST]] into [[IR]].
-    *
-    * @param arg the argument to translate
-    * @param isSuspended `true` if the argument is suspended, otherwise `false`
-    * @return the [[IR]] representation of `arg`
-    * @tailrec
-    */
-  IR.DefinitionArgument translateArgumentDefinition(Tree arg, boolean isSuspended) {
-    var core = maybeManyParensed(arg);
-    return switch (core) {
-      case null -> null;
-      case Tree.OprApp app when isOperator(":", app.getOpr()) -> {
-        yield switch (translateIdent(app.getLhs(), false)) {
-          case IR.Name name -> {
-            var type = translateQualifiedNameOrExpression(app.getRhs());
-            yield new IR$DefinitionArgument$Specified(
-              name,
-              Option.apply(type), Option.empty(),
-              false, getIdentifiedLocation(app), meta(), diag()
-            );
-          }
-          default -> throw new UnhandledEntity(app.getLhs(), "translateArgumentDefinition");
-        };
-      }
-      case Tree.OprApp withValue when isOperator("=", withValue.getOpr()) -> {
-        var defaultValue = translateExpression(withValue.getRhs(), false);
-        yield switch (withValue.getLhs()) {
-          case Tree.OprApp app when isOperator(":", app.getOpr()) -> {
-            yield switch (translateIdent(app.getLhs(), false)) {
-              case IR.Name name -> {
-                var type = translateQualifiedNameOrExpression(app.getRhs());
-                yield new IR$DefinitionArgument$Specified(
-                  name,
-                  Option.apply(type), Option.apply(defaultValue),
-                  false, getIdentifiedLocation(app), meta(), diag()
-                );
-              }
-              default -> throw new UnhandledEntity(app.getLhs(), "translateArgumentDefinition");
-            };
-          }
-          case Tree.TypeAnnotated anno -> {
-            yield null;
-          }
-          default -> throw new UnhandledEntity(withValue.getLhs(), "translateArgumentDefinition");
-        };
-      }
-      case Tree.OprSectionBoundary bound -> {
-          yield translateArgumentDefinition(bound.getAst(), isSuspended);
-      }
-      case Tree.TypeAnnotated anno -> {
-        yield null;
-      }
-      /*
-      case AstView.AscribedArgument(name, ascType, mValue, isSuspended) =>
-        translateIdent(name) match {
-          case name: IR.Name =>
-            DefinitionArgument.Specified(
-              name,
-              Some(translateQualifiedNameOrExpression(ascType)),
-              mValue.map(translateExpression(_)),
-              isSuspended,
-              getIdentifiedLocation(arg)
-            )
-          case _ =>
-            throw new UnhandledEntity(arg, "translateArgumentDefinition")
-        }
-      case AstView.LazyAssignedArgumentDefinition(name, value) =>
-        translateIdent(name) match {
-          case name: IR.Name =>
-            DefinitionArgument.Specified(
-              name,
-              None,
-              Some(translateExpression(value)),
-              suspended = true,
-              getIdentifiedLocation(arg)
-            )
-          case _ =>
-            throw new UnhandledEntity(arg, "translateArgumentDefinition")
-        }
-      case AstView.LazyArgument(arg) =>
-        translateArgumentDefinition(arg, isSuspended = true)
-        */
+   *
+   * @param def the argument to translate
+   * @return the [[IR]] representation of `arg`
+   */
+  IR.DefinitionArgument translateArgumentDefinition(ArgumentDefinition def) {
+    Tree pattern = def.getPattern();
+    IR.Name name = switch (pattern) {
+      case Tree.Wildcard wild -> new IR$Name$Blank(getIdentifiedLocation(wild.getToken()), meta(), diag());
       case Tree.Ident id -> {
         IR.Expression identifier = translateIdent(id, false);
         yield switch (identifier) {
-          case IR.Name name -> new IR$DefinitionArgument$Specified(
-            name,
-            Option.empty(),
-            Option.empty(),
-            isSuspended,
-            getIdentifiedLocation(arg),
-            meta(), diag()
-          );
-          default -> throw new UnhandledEntity(arg, "translateArgumentDefinition");
+          case IR.Name name_ -> name_;
+          // TODO: Other types of pattern. Needs IR support.
+          default -> throw new UnhandledEntity(pattern, "translateArgumentDefinition");
         };
       }
-      case Tree.UnaryOprApp opr when "~".equals(opr.getOpr().codeRepr()) -> {
-        yield translateArgumentDefinition(opr.getRhs(), true);
-      }
-        /*
-      case AstView.AssignedArgument(name, value) =>
-        translateIdent(name) match {
-          case name: IR.Name =>
-            DefinitionArgument.Specified(
-              name,
-              None,
-              Some(translateExpression(value)),
-              isSuspended,
-              getIdentifiedLocation(arg)
-            )
-          case _ =>
-            throw new UnhandledEntity(arg, "translateArgumentDefinition")
-        }
-      */
-      case Tree.Wildcard wild -> {
-        var loc = getIdentifiedLocation(arg);
-        yield new IR$DefinitionArgument$Specified(
-            new IR$Name$Blank(loc, meta(), diag()),
-            Option.empty(),
-            Option.empty(),
-            isSuspended,
-            loc,
-            meta(), diag()
-          );
-      }
-      default -> throw new UnhandledEntity(core, "translateArgumentDefinition");
+      // TODO: Other types of pattern. Needs IR support.
+      default -> throw new UnhandledEntity(pattern, "translateArgumentDefinition");
     };
+    boolean isSuspended = def.getSuspension() != null;
+    var ascribedType = Option.apply(def.getType()).map(ascription -> translateExpression(ascription.getType(), true));
+    var defaultValue = Option.apply(def.getDefault()).map(default_ -> translateExpression(default_.getExpression(), false));
+    return new IR$DefinitionArgument$Specified(
+            name,
+            ascribedType,
+            defaultValue,
+            isSuspended,
+            getIdentifiedLocation(def),
+            meta(),
+            diag()
+    );
   }
 
+  @SuppressWarnings("unchecked")
   private List<IR.CallArgument> translateCallArguments(List<Tree> args, List<IR.CallArgument> res, boolean insideTypeSignature) {
     while (args.nonEmpty()) {
       var a = translateCallArgument(args.head(), insideTypeSignature);
@@ -1210,21 +1213,19 @@ final class TreeToIr {
     Tree arg,
     boolean insideTypeSignature
   ) {
-    /*
-    arg match {
-      case AstView.AssignedArgument(left, right) =>
-        CallArgument
-          .Specified(
-            Some(buildName(left)),
-            translateExpression(right, insideTypeSignature),
-            getIdentifiedLocation(arg)
-          )
-      case _ =>
-    }
-    */
-    var expr = translateExpression(arg, insideTypeSignature);
     var loc = getIdentifiedLocation(arg);
-    return new IR$CallArgument$Specified(Option.empty(), expr, loc, meta(), diag());
+    return switch (arg) {
+      case Tree.NamedApp app -> {
+        var expr = translateExpression(app.getArg(), insideTypeSignature);
+        var id = buildName(app, app.getName());
+        yield new IR$CallArgument$Specified(Option.apply(id), expr, loc, meta(), diag());
+      }
+      case null -> null;
+      default -> {
+        var expr = translateExpression(arg, insideTypeSignature);
+        yield new IR$CallArgument$Specified(Option.empty(), expr, loc, meta(), diag());
+      }
+    };
   }
 
   /** Calculates whether a set of arguments has its defaults suspended, and
@@ -1458,54 +1459,84 @@ final class TreeToIr {
     *
     * @param branch the case branch or comment to translate
     * @return the [[IR]] representation of `branch`
-
-  def translateCaseBranch(branch: AST): Case.Branch = {
-    branch match {
-      case AstView.CaseBranch(pattern, expression) =>
-        Case.Branch(
-          translatePattern(pattern),
-          translateExpression(expression),
-          getIdentifiedLocation(branch)
-        )
-      case c @ AST.Comment(lines) =>
-        val doc      = lines.mkString("\n")
-        val location = getIdentifiedLocation(c)
-        Case.Branch(
-          Pattern.Documentation(doc, location),
-          IR.Empty(None),
-          location
-        )
-      case _ => throw new UnhandledEntity(branch, "translateCaseBranch")
+    */
+  IR$Case$Branch translateCaseBranch(Case branch) {
+    if (branch.getPattern() == null) {
+      throw new UnhandledEntity(branch, "translateCaseBranch");
     }
+    return new IR$Case$Branch(
+          translatePattern(branch.getPattern(), nil()),
+          translateExpression(branch.getExpression(), false),
+          getIdentifiedLocation(branch.getExpression()), meta(), diag()
+      );
   }
 
   /** Translates a pattern in a case expression from its [[AST]] representation
     * into [[IR]].
     *
-    * @param pattern the case pattern to translate
+    * @param block the case pattern to translate
     * @return
+    */
+  IR.Pattern translatePattern(Tree block, List<IR.Pattern> fields) {
+    var pattern = maybeManyParensed(block);
+    return switch (pattern) {
+      case Tree.Ident id -> {
+        yield new IR$Pattern$Constructor(
+          buildName(id), fields,
+          getIdentifiedLocation(id), meta(), diag()
+        );
+      }
+      case Tree.OprApp id -> {
+        var qualifiedName = buildQualifiedName(pattern);
+        yield new IR$Pattern$Constructor(
+          qualifiedName, fields,
+          getIdentifiedLocation(id), meta(), diag()
+        );
+      }
+      case Tree.App id -> {
+        var args = translatePatternArguments(id.getArg(), fields);
+        yield translatePattern(id.getFunc(), args);
+      }
+      case Tree.Wildcard wild -> translateWildcardPattern(wild);
+      case Tree.TextLiteral lit -> {
+        yield new IR$Pattern$Literal(translateLiteral(lit), getIdentifiedLocation(lit), meta(), diag());
+      }
+      case Tree.Number num -> {
+        yield new IR$Pattern$Literal((IR.Literal) translateDecimalLiteral(num), getIdentifiedLocation(num), meta(), diag());
+      }
+      default -> throw new UnhandledEntity(pattern, "translatePattern");
+    };
+  }
 
-  def translatePattern(pattern: AST): Pattern = {
-    AstView.MaybeManyParensed.unapply(pattern).getOrElse(pattern) match {
-      case AstView.ConstructorPattern(conses, fields) =>
-        val irConses = conses.map(translateIdent(_).asInstanceOf[IR.Name])
-        val name = irConses match {
-          case List(n) => n
-          case _       => IR.Name.Qualified(irConses, None)
-        }
-        Pattern.Constructor(
-          name,
-          fields.map(translatePattern),
-          getIdentifiedLocation(pattern)
-        )
-      case AstView.CatchAllPattern(name) =>
-        Pattern.Name(
-          translateIdent(name).asInstanceOf[IR.Name],
-          getIdentifiedLocation(pattern)
-        )
-      case _ =>
-        throw new UnhandledEntity(pattern, "translatePattern")
-    }
+  private List<IR.Pattern> translatePatternArguments(Tree t, List<IR.Pattern> prev) {
+    var tree = maybeManyParensed(t);
+    return switch (tree) {
+      case Tree.OprApp app -> {
+       var tail = translatePatternArguments(app.getRhs(), prev);
+       yield cons(translatePattern(app.getLhs(), tail), nil());
+      }
+      case Tree.App app -> {
+       var tail = translatePatternArguments(app.getArg(), prev);
+       yield cons(translatePattern(app.getFunc(), tail), nil());
+      }
+      case Tree.Ident id -> {
+        var name = buildName(id);
+        var pattern = Character.isLowerCase(name.name().charAt(0)) ?
+                new IR$Pattern$Name(name, getIdentifiedLocation(id), meta(), diag()) :
+                new IR$Pattern$Constructor(name, nil(), getIdentifiedLocation(id), meta(), diag());
+        yield cons(pattern, prev);
+      }
+      case Tree.Wildcard wild -> cons(translateWildcardPattern(wild), prev);
+      default -> throw new UnhandledEntity(tree, "translatePattern");
+    };
+  }
+
+  private IR$Pattern$Name translateWildcardPattern(Tree.Wildcard wild) {
+      var at = getIdentifiedLocation(wild);
+      var blank = new IR$Name$Blank(
+              at, meta(), diag()
+      );
+      return new IR$Pattern$Name(blank, at, meta(), diag());
   }
 
   /** Translates an arbitrary grouped piece of syntax from its [[AST]]
@@ -1565,61 +1596,30 @@ final class TreeToIr {
     * @return the [[IR]] representation of `imp`
     */
   IR$Module$Scope$Import translateImport(Tree.Import imp) {
-    if (imp.getImport() != null) {
-      if (imp.getFrom() != null) {
-        var qualifiedName = buildQualifiedName(imp.getFrom().getBody());
-        var onlyNames = imp.getImport().getBody();
-        var isAll = isAll(onlyNames);
-        return new IR$Module$Scope$Import$Module(
-          qualifiedName, Option.empty(), isAll, Option.empty(),
-          Option.empty(), getIdentifiedLocation(imp), false,
-          meta(), diag()
-        );
-      } else if (imp.getPolyglot() != null) {
-        List<IR.Name> qualifiedName = buildNames(imp.getImport().getBody(), '.', true);
-        StringBuilder pkg = new StringBuilder();
-        String cls = extractPackageAndName(qualifiedName, pkg);
-        Option<String> rename = imp.getImportAs() == null ? Option.empty() :
-                Option.apply(buildName(imp.getImportAs().getBody()).name());
-        return new IR$Module$Scope$Import$Polyglot(
-          new IR$Module$Scope$Import$Polyglot$Java(pkg.toString(), cls),
-          rename, getIdentifiedLocation(imp),
-          meta(), diag()
-        );
-      } else {
-        var qualifiedName = buildQualifiedName(imp.getImport().getBody());
-        Option<IR$Name$Literal> rename = imp.getImportAs() == null ? Option.empty() :
-                Option.apply(buildName(imp.getImportAs().getBody()));
-        return new IR$Module$Scope$Import$Module(
-          qualifiedName, rename, false, Option.empty(),
-          Option.empty(), getIdentifiedLocation(imp), false,
-          meta(), diag()
-        );
-      }
+    Option<IR$Name$Literal> rename = Option.apply(imp.getAs()).map(as -> buildName(as.getBody()));
+    if (imp.getPolyglot() != null) {
+      List<IR.Name> qualifiedName = buildNames(imp.getImport().getBody(), '.', true);
+      StringBuilder pkg = new StringBuilder();
+      String cls = extractPackageAndName(qualifiedName, pkg);
+      return new IR$Module$Scope$Import$Polyglot(
+              new IR$Module$Scope$Import$Polyglot$Java(pkg.toString(), cls),
+              rename.map(name -> name.name()), getIdentifiedLocation(imp),
+              meta(), diag()
+      );
     }
-    /*
-    imp match {
-      case AST.Import(path, rename, isAll, onlyNames, hiddenNames) =>
-        IR.Module.Scope.Import.Module(
-          IR.Name.Qualified(path.map(buildName(_)).toList, None),
-          rename.map(buildName(_)),
-          isAll,
-          onlyNames.map(_.map(buildName(_)).toList),
-          hiddenNames.map(_.map(buildName(_)).toList),
-          getIdentifiedLocation(imp)
-        )
-      case _ =>
-        IR.Error.Syntax(imp, IR.Error.Syntax.InvalidImport)
+    IR$Name$Qualified qualifiedName;
+    // TODO: onlyNames, hiddenNames
+    if (imp.getFrom() != null) {
+      qualifiedName = buildQualifiedName(imp.getFrom().getBody());
+    } else {
+      qualifiedName = buildQualifiedName(imp.getImport().getBody());
     }
-    */
-    return new IR$Error$Syntax(imp, IR$Error$Syntax$InvalidImport$.MODULE$, meta(), diag());
-  }
-
-  private boolean isAll(Tree onlyNames) {
-    return switch (onlyNames) {
-      case Tree.Ident id -> buildName(id).name().equals("all");
-      default -> false;
-    };
+    var isAll = imp.getAll() != null;
+    return new IR$Module$Scope$Import$Module(
+      qualifiedName, rename, isAll, Option.empty(),
+      Option.empty(), getIdentifiedLocation(imp), false,
+      meta(), diag()
+    );
   }
 
   @SuppressWarnings("unchecked")
@@ -1644,56 +1644,33 @@ final class TreeToIr {
     * @param exp the export to translate
     * @return the [[IR]] representation of `imp`
     */
+  @SuppressWarnings("unchecked")
   IR$Module$Scope$Export$Module translateExport(Tree.Export exp) {
-    if (exp.getExport() != null) {
-      if (exp.getFrom() != null) {
-        var qualifiedName = buildQualifiedName(exp.getFrom().getBody());
-        var onlyBodies = exp.getExport().getBody();
-        var isAll = isAll(onlyBodies);
-        final Option<List<IR$Name$Literal>> onlyNames = isAll ? Option.empty() :
-          Option.apply((List<IR$Name$Literal>) (Object)buildNames(onlyBodies, ',', false));
+    Option<IR$Name$Literal> rename = Option.apply(exp.getAs()).map(as -> buildName(as.getBody()));
+    if (exp.getFrom() != null) {
+      var qualifiedName = buildQualifiedName(exp.getFrom().getBody());
+      var onlyBodies = exp.getExport().getBody();
+      var isAll = exp.getAll() != null;
+      final Option<List<IR$Name$Literal>> onlyNames = isAll ? Option.empty() :
+        Option.apply((List<IR$Name$Literal>) (Object)buildNames(onlyBodies, ',', false));
 
-        var hidingList = exp.getHiding() == null ? nil() : buildNames(exp.getHiding().getBody(), ',', false);
-        final Option<List<IR$Name$Literal>> hidingNames = hidingList.isEmpty() ? Option.empty() :
-          Option.apply((List<IR$Name$Literal>) (Object)hidingList);
+      var hidingList = exp.getHiding() == null ? nil() : buildNames(exp.getHiding().getBody(), ',', false);
+      final Option<List<IR$Name$Literal>> hidingNames = hidingList.isEmpty() ? Option.empty() :
+        Option.apply((List<IR$Name$Literal>) (Object)hidingList);
 
-        Option<IR$Name$Literal> rename;
-        if (exp.getFromAs() != null) {
-          rename = Option.apply(buildName(exp.getFromAs().getBody()));
-        } else {
-          rename = Option.empty();
-        }
-
-        return new IR$Module$Scope$Export$Module(
-          qualifiedName, rename,
-          true, onlyNames, hidingNames, getIdentifiedLocation(exp), false,
-          meta(), diag()
+      return new IR$Module$Scope$Export$Module(
+        qualifiedName, rename,
+        true, onlyNames, hidingNames, getIdentifiedLocation(exp), false,
+        meta(), diag()
+      );
+    } else {
+      var qualifiedName = buildQualifiedName(exp.getExport().getBody());
+      return new IR$Module$Scope$Export$Module(
+        qualifiedName, rename, false, Option.empty(),
+        Option.empty(), getIdentifiedLocation(exp), false,
+        meta(), diag()
         );
-      } else {
-        var qualifiedName = buildQualifiedName(exp.getExport().getBody());
-        Option<IR$Name$Literal> rename = exp.getExportAs() == null ? Option.empty() :
-                Option.apply(buildName(exp.getExportAs().getBody()));
-        return new IR$Module$Scope$Export$Module(
-          qualifiedName, rename, false, Option.empty(),
-          Option.empty(), getIdentifiedLocation(exp), false,
-          meta(), diag()
-        );
-      }
     }
-    /*
-    exp match {
-      case AST.Export(path, rename, isAll, onlyNames, hiddenNames) =>
-        IR.Module.Scope.Export.Module(
-          IR.Name.Qualified(path.map(buildName(_)).toList, None),
-          rename.map(buildName(_)),
-          isAll,
-          onlyNames.map(_.map(buildName(_)).toList),
-          hiddenNames.map(_.map(buildName(_)).toList),
-          getIdentifiedLocation(exp)
-        )
-    }
-      case _ -> */
-    throw new UnhandledEntity(exp, "translateExport");
   }
 
   /** Translates an arbitrary invalid expression from the [[AST]] representation
@@ -1737,7 +1714,7 @@ final class TreeToIr {
     *
     * @param tree the comment to transform
     * @return the [[IR]] representation of `comment`
-    *
+    */
   IR.Comment translateComment(Tree tree) {
     return switch (tree) {
       case Tree.ArgumentBlockApplication comment -> {
@@ -1756,11 +1733,18 @@ final class TreeToIr {
         throw new UnhandledEntity(tree, "translateComment");
     };
   }
-  */
 
+  private IR$Name$Literal buildName(Token name) {
+    return new IR$Name$Literal(
+            name.codeRepr(),
+            false,
+            getIdentifiedLocation(name),
+            meta(), diag()
+    );
+  }
   private IR$Name$Literal buildName(Tree ident) {
     return switch (ident) {
-      case Tree.Ident id -> buildName(id, id.getToken());
+      case Tree.Ident id -> buildName(id.getToken());
       default -> throw new UnhandledEntity(ident, "buildName");
     };
   }
@@ -1777,16 +1761,52 @@ final class TreeToIr {
   }
 
   private Option<IdentifiedLocation> getIdentifiedLocation(Tree ast) {
-    if (ast == null) {
-      return Option.empty();
+    return Option.apply(ast).map(ast_ -> {
+      int begin = Math.toIntExact(ast_.getStartCode());
+      int end = Math.toIntExact(ast_.getEndCode());
+      return new IdentifiedLocation(new Location(begin, end), Option.empty());
+    });
+  }
+
+  private Option<IdentifiedLocation> getIdentifiedLocation(ArgumentDefinition ast) {
+    // Note: In the IR, `DefinitionArgument` is a type of AST node; in the parser, it is not an AST-level type, but a
+    // type used only in specific locations. As a result, the IR expects it to have a source-code location, but the
+    // parser doesn't have a span for it. Here we synthesize one. This will not be necessary if we refactor IR so that
+    // `ArgumentDefinition` is not an AST node, though that change would have effects throughout the compiler and may
+    // not be worthwhile if IR is expected to be replaced.
+    long begin;
+    if (ast.getOpen() != null) {
+      begin = ast.getOpen().getStartCode();
+    } else if (ast.getOpen2() != null) {
+      begin = ast.getOpen2().getStartCode();
+    } else if (ast.getSuspension() != null) {
+      begin = ast.getSuspension().getStartCode();
+    } else {
+      begin = ast.getPattern().getStartCode();
     }
-    int begin, len;
-    {
-      begin = Math.toIntExact(ast.getStartCode());
-      int end = Math.toIntExact(ast.getEndCode());
-      len = end - begin;
+    int begin_ = Math.toIntExact(begin);
+    long end;
+    if (ast.getClose() != null) {
+      end = ast.getClose().getEndCode();
+    } else if (ast.getDefault() != null) {
+      end = ast.getDefault().getEquals().getEndCode();
+    } else if (ast.getClose2() != null) {
+      end = ast.getClose2().getEndCode();
+    } else if (ast.getType() != null) {
+      end = ast.getType().getOperator().getEndCode();
+    } else {
+      end = ast.getPattern().getEndCode();
     }
-    return Option.apply(new IdentifiedLocation(new Location(begin, begin + len), Option.empty()));
+    int end_ = Math.toIntExact(end);
+    return Option.apply(new IdentifiedLocation(new Location(begin_, end_), Option.empty()));
+  }
+
+  private Option<IdentifiedLocation> getIdentifiedLocation(Token ast) {
+    return Option.apply(ast).map(ast_ -> {
+      int begin = Math.toIntExact(ast_.getStartCode());
+      int end = Math.toIntExact(ast_.getEndCode());
+      return new IdentifiedLocation(new Location(begin, end), Option.empty());
+    });
   }
   private MetadataStorage meta() {
     return MetadataStorage.apply(nil());