+++ lib/rust/parser/src/syntax/item.rs
@@ -44,92 +44,7 @@ impl<'s> Item<'s> {
     /// Convert this item to a [`Tree`].
     pub fn to_ast(self) -> Tree<'s> {
         match self {
-            Item::Token(token) => match token.variant {
-                token::Variant::Ident(ident) => Tree::ident(token.with_variant(ident)),
-                token::Variant::Digits(number) =>
-                    Tree::number(None, Some(token.with_variant(number)), None),
-                token::Variant::NumberBase(base) =>
-                    Tree::number(Some(token.with_variant(base)), None, None),
-                token::Variant::TextStart(open) => Tree::text_literal(
-                    Some(token.with_variant(open)),
-                    default(),
-                    default(),
-                    default(),
-                ),
-                token::Variant::TextSection(section) => {
-                    let trim = token.left_offset.visible;
-                    let section = tree::TextElement::Section { text: token.with_variant(section) };
-                    Tree::text_literal(default(), vec![section], default(), trim)
-                }
-                token::Variant::TextEscapeSymbol(escape) => {
-                    let trim = token.left_offset.visible;
-                    let backslash = Some(token.with_variant(escape));
-                    let section = tree::TextElement::EscapeChar { backslash, char: None };
-                    Tree::text_literal(default(), vec![section], default(), trim)
-                }
-                token::Variant::TextEscapeChar(escape) => {
-                    let trim = token.left_offset.visible;
-                    let char = Some(token.with_variant(escape));
-                    let section = tree::TextElement::EscapeChar { backslash: None, char };
-                    Tree::text_literal(default(), vec![section], default(), trim)
-                }
-                token::Variant::TextEscapeLeader(leader) => {
-                    let trim = token.left_offset.visible;
-                    let leader = Some(token.with_variant(leader));
-                    let section = tree::TextElement::EscapeSequence {
-                        leader,
-                        open: None,
-                        digits: None,
-                        close: None,
-                    };
-                    Tree::text_literal(default(), vec![section], default(), trim)
-                }
-                token::Variant::TextEscapeHexDigits(digits) => {
-                    let digits = Some(token.with_variant(digits));
-                    let section = tree::TextElement::EscapeSequence {
-                        leader: None,
-                        open: None,
-                        digits,
-                        close: None,
-                    };
-                    Tree::text_literal(default(), vec![section], default(), default())
-                }
-                token::Variant::TextEscapeSequenceStart(t) => {
-                    let open = Some(token.with_variant(t));
-                    let section = tree::TextElement::EscapeSequence {
-                        leader: None,
-                        open,
-                        digits: None,
-                        close: None,
-                    };
-                    Tree::text_literal(default(), vec![section], default(), default())
-                }
-                token::Variant::TextEscapeSequenceEnd(t) => {
-                    let close = Some(token.with_variant(t));
-                    let section = tree::TextElement::EscapeSequence {
-                        leader: None,
-                        open: None,
-                        digits: None,
-                        close,
-                    };
-                    Tree::text_literal(default(), vec![section], default(), default())
-                }
-                token::Variant::TextEnd(close) => Tree::text_literal(
-                    default(),
-                    default(),
-                    Some(token.with_variant(close)),
-                    default(),
-                ),
-                token::Variant::Wildcard(wildcard) => Tree::wildcard(token.with_variant(wildcard)),
-                token::Variant::AutoScope(t) => Tree::auto_scope(token.with_variant(t)),
-                _ => {
-                    let message = format!("to_ast: Item::Token({token:?})");
-                    let value = Tree::ident(
-                        token.with_variant(token::variant::Ident(false, 0, false, false)),
-                    );
-                    Tree::with_unsupported(value, message)
-                }
-            },
+            Item::Token(token) => token.into(),
             Item::Tree(ast) => ast,
             Item::Block(items) => build_block(items),
         }
@@ -202,7 +117,9 @@ macro_rules! generate_variant_checks {
         pub enum $enum:ident {
             $(
                 $(#$variant_meta:tt)*
-                $variant:ident $({ $(pub $field:ident : $field_ty:ty),* $(,)? })?
+                $variant:ident $({
+                    $($(#$field_meta:tt)* pub $field:ident : $field_ty:ty),* $(,)?
+                })?
             ),* $(,)?
         }
     ) => { paste!{